rebase copy_image implementation
