+++ src/cmake/testing.cmake
@@ -274,7 +274,7 @@ macro (oiio_add_all_tests)
                     IMAGEDIR j2kp4files_v1_5
                     URL http://www.itu.int/net/ITU-T/sigdb/speimage/ImageForm-s.aspx?val=10100803)
     set (all_openexr_tests
-         openexr-suite openexr-multires openexr-chroma
+         openexr-suite openexr-multires openexr-chroma openexr-copy
          openexr-v2 openexr-window perchannel oiiotool-deep)
     if (OpenEXR_VERSION VERSION_GREATER_EQUAL 3.1.10)
         # OpenEXR 3.1.10 is the first release where the exr core library
